+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Added
+  - Experimental support for downloading F2D derivatives
+
 ## [1.4.8] - 2019-12-11
 
 - Fixed